+++ src/p_user.c
@@ -395,7 +395,7 @@ void P_PlayerThink(player_t *player)
     }
 
     // [BH] regenerate health up to 100 every 1 second
-    if (player->mo->health < initial_health && (player->cheats & CF_BUDDHA)
+    if (player->mo->health < initial_health && (regenhealth || (player->cheats & CF_BUDDHA))
         && !(leveltime % TICRATE) && !player->damagecount)
         player->mo->health = player->health = MIN(player->health + 1, initial_health);